% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_sim.R
\name{all_sim}
\alias{all_sim}
\title{Cosine similarty between two users in a recommender system.}
\usage{
all_sim(a, b, mat)
}
\arguments{
\item{a}{User ID of first user.}

\item{b}{User ID of second user.}

\item{mat}{Spread matrix in the form of users, items, ratings.}
}
\value{
Returns the cosine similarity of two users.
}
\description{
This function isolates two users in a recommender system and returns their cosine similarity.
}
