% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_sim.R
\name{cosine_sim}
\alias{cosine_sim}
\title{Cosine Similarity}
\usage{
cosine_sim(a, b)
}
\arguments{
\item{a}{Vector one}

\item{b}{Vector two}
}
\value{
Returns the cosine similarity value between two vectors
}
\description{
This function returns the cosine similarity value between two vectors.
}
