% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_set.R
\name{get_test_set}
\alias{get_test_set}
\title{Get Test Set}
\usage{
get_test_set(ratings, user_col, item_col, rating_col)
}
\arguments{
\item{ratings}{A dataset of users, items and ratings that has been removed of zero ratings.}

\item{user_col}{The name of the user ID column in the ratings dataframe.}

\item{item_col}{The name of the item ID column in the ratings dataframe.}

\item{rating_col}{The name of the ratings column in the ratings dataframe.}
}
\value{
A matrix in the form of the User ID, index of the user's first rating that was set to NA,
index of the user's second rating that was set to NA, original value of the user's first rating
and original value of the user's second rating.
}
\description{
This function is used to create a test/validation set for a matrix decomposition recommender system.
}
\examples{
get_test_set(book_ratings,'User.ID', 'ISBN', 'Book.Rating')
}
