% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_based_rating.R
\name{item_based_rating}
\alias{item_based_rating}
\title{Item Based Rating}
\usage{
item_based_rating(user_id, ratings, items, similarities)
}
\arguments{
\item{user_id}{The ID of the user.}

\item{ratings}{Spread matrix in the form of users (rows), items (columns) and ratings (values).}

\item{items}{Dataframe of items and information on the items.}

\item{similarities}{Similarity matrix of items in the system.}
}
\value{
Returns a sorted dataframe of the top 10 rated items, the score of the item and information on each item.
}
\description{
This funciton returns the top 10 rated item rated items of any user in already in the dataset used to create the recommender system.
}
