% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_decomp.R
\name{matrix_decomp}
\alias{matrix_decomp}
\title{Matrix Decomposition}
\usage{
matrix_decomp(ratings, changed, k_in, user_col, item_col, rating_col, a, b)
}
\arguments{
\item{ratings}{A dataset of users, items and ratings that has been removed of zero ratings.}

\item{changed}{A matrix in the form of the User ID, index of the user's first rating that was set to NA,
index of the user's second rating that was set to NA, original value of the user's first rating
and original value of the user's second rating.}

\item{user_col}{The name of the user ID column in the ratings dataframe.}

\item{item_col}{The name of the item ID column in the ratings dataframe.}

\item{rating_col}{The name of the ratings column in the ratings dataframe.}

\item{a}{The L2 regularisation of W latent factors for the matrix decomposition.}

\item{b}{The L2 regularisation of H latent factors for the matrix decomposition}

\item{k}{The integer of decomposition rank of the decomposition matrix.}
}
\value{
This functions returns both the predicted ratings and RMSE of a matrix decomposition recommender system.
}
\description{
This functions returns the predicted ratings and RMSE for a matrix decomposition recommender system.
}
\examples{
matrix_decomp(book_ratings, changed_inds,1,'User.ID', 'ISBN', 'Book.Rating', 0, 0)
}
