% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_to_zero.R
\name{na_to_zero}
\alias{na_to_zero}
\title{NA Values to Zero}
\usage{
na_to_zero(mat)
}
\arguments{
\item{mat}{A spread matrix in the from users, items and ratings}
}
\value{
Returns the parameter spread matrix where all na values have been set to zero
}
\description{
This function sets all na zalues in a spread matrix to zero
}
\examples{
na_to_zero(book_ratings_matrix)
}
