% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_zero_ratings.R
\name{remove_zero_ratings}
\alias{remove_zero_ratings}
\title{Remove Zero Ratings}
\usage{
remove_zero_ratings(frame, rating)
}
\arguments{
\item{frame}{This is a dataframe containing users, items and ratings}

\item{rating}{Column name of the column containing the ratings.}
}
\value{
Dataframe with zero ratings removed
}
\description{
Removes zero ratings from a dataframe containing users, items, and ratings.
}
\examples{
remove_zero_ratings(book_ratings, 'Book.Rating')
}
