% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_matrix.R
\name{sim_matrix}
\alias{sim_matrix}
\title{Item Similarity Matrix}
\usage{
sim_matrix(ma, mb)
}
\arguments{
\item{ma}{The transpose of the spread matrix created for the recommender system. The rows of the spread matrix represent users, the columns represent items and the values represent ratings.}

\item{mb}{The same parameter as above.}
}
\value{
Similarity matrix between the items in a an item based remommender system.
}
\description{
This function returns a similarity matrix for items in an item based remommender system.
}
\examples{
sim_matrix(t(book_ratings_matrix), t(book_ratings_matrix))
}
